refactor(api): deprecate PATCH /organizations/{id}/schooling-registration-user-associations/{schoolingRegistrationId}
in favor of PATCH /organizations/{id}/sup-organization-learners/{organizationLearnerId}
