--- api/lib/application/organization-learner-user-associations/index.js
@@ -1,53 +1,0 @@
-const Joi = require('joi').extend(require('@joi/date'));
-
-const { sendJsonApiError, UnprocessableEntityError, NotFoundError } = require('../http-errors');
-const securityPreHandlers = require('../security-pre-handlers');
-const organizationLearnerUserAssociationController = require('./organization-learner-user-association-controller');
-const identifiersType = require('../../domain/types/identifiers-type');
-
-exports.register = async function (server) {
-  server.route([
-    {
-      method: 'PATCH',
-      path: '/api/organizations/{id}/schooling-registration-user-associations/{schoolingRegistrationId}',
-      config: {
-        pre: [
-          {
-            method: securityPreHandlers.checkUserIsAdminInSUPOrganizationManagingStudents,
-          },
-        ],
-        handler: organizationLearnerUserAssociationController.updateStudentNumber,
-        validate: {
-          options: {
-            allowUnknown: true,
-          },
-          params: Joi.object({
-            id: identifiersType.organizationId,
-            schoolingRegistrationId: identifiersType.schoolingRegistrationId,
-          }),
-          payload: Joi.object({
-            data: {
-              attributes: {
-                'student-number': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
-              },
-            },
-          }),
-          failAction: (request, h, err) => {
-            const isStudentNumber = err.details[0].path.includes('student-number');
-            if (isStudentNumber) {
-              return sendJsonApiError(new UnprocessableEntityError('Un des champs saisis n’est pas valide.'), h);
-            }
-            return sendJsonApiError(new NotFoundError('Ressource non trouvée'), h);
-          },
-        },
-        notes: [
-          "- **Cette route est restreinte aux utilisateurs authentifiés et admin au sein de l'orga**\n" +
-            '- Elle met à jour le numéro étudiant',
-        ],
-        tags: ['api', 'organizationLearnerUserAssociation'],
-      },
-    },
-  ]);
-};
-
-exports.name = 'schooling-registration-user-associations-api';
