+++ api/lib/application/sup-organization-learners/index.js
@@ -1,7 +1,9 @@
 const Joi = require('joi').extend(require('@joi/date'));
 
-const { sendJsonApiError, UnprocessableEntityError } = require('../http-errors');
+const { sendJsonApiError, UnprocessableEntityError, NotFoundError } = require('../http-errors');
 const supOrganizationLearnerController = require('./sup-organization-learner-controller');
+const securityPreHandlers = require('../security-pre-handlers');
+const identifiersType = require('../../domain/types/identifiers-type');
 
 exports.register = async function (server) {
   server.route([
@@ -68,6 +70,87 @@ exports.register = async function (server) {
             '- Elle réconcilie l’utilisateur à l’inscription d’un étudiant dans cette organisation',
         ],
         tags: ['api', 'sup-organization-learner'],
+      },
+    },
+    {
+      method: 'PATCH',
+      path: '/api/organizations/{id}/schooling-registration-user-associations/{schoolingRegistrationId}',
+      config: {
+        pre: [
+          {
+            method: securityPreHandlers.checkUserIsAdminInSUPOrganizationManagingStudents,
+          },
+        ],
+        handler: supOrganizationLearnerController.updateStudentNumber,
+        validate: {
+          options: {
+            allowUnknown: true,
+          },
+          params: Joi.object({
+            id: identifiersType.organizationId,
+            schoolingRegistrationId: identifiersType.schoolingRegistrationId,
+          }),
+          payload: Joi.object({
+            data: {
+              attributes: {
+                'student-number': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
+              },
+            },
+          }),
+          failAction: (request, h, err) => {
+            const isStudentNumber = err.details[0].path.includes('student-number');
+            if (isStudentNumber) {
+              return sendJsonApiError(new UnprocessableEntityError('Un des champs saisis n’est pas valide.'), h);
+            }
+            return sendJsonApiError(new NotFoundError('Ressource non trouvée'), h);
+          },
+        },
+        notes: [
+          "- **Cette route est restreinte aux utilisateurs authentifiés et admin au sein de l'orga**\n" +
+            '- Elle met à jour le numéro étudiant',
+          "- L'usage de cette route est **dépréciée** en faveur de /api/organizations/{id}/sup-organization-learners/{organizationLearnerId}",
+        ],
+        tags: ['api', 'organizationLearnerUserAssociation'],
+      },
+    },
+    {
+      method: 'PATCH',
+      path: '/api/organizations/{id}/sup-organization-learners/{organizationLearnerId}',
+      config: {
+        pre: [
+          {
+            method: securityPreHandlers.checkUserIsAdminInSUPOrganizationManagingStudents,
+          },
+        ],
+        handler: supOrganizationLearnerController.updateStudentNumber,
+        validate: {
+          options: {
+            allowUnknown: true,
+          },
+          params: Joi.object({
+            id: identifiersType.organizationId,
+            organizationLearnerId: identifiersType.organizationLearnerId,
+          }),
+          payload: Joi.object({
+            data: {
+              attributes: {
+                'student-number': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
+              },
+            },
+          }),
+          failAction: (request, h, err) => {
+            const isStudentNumber = err.details[0].path.includes('student-number');
+            if (isStudentNumber) {
+              return sendJsonApiError(new UnprocessableEntityError('Un des champs saisis n’est pas valide.'), h);
+            }
+            return sendJsonApiError(new NotFoundError('Ressource non trouvée'), h);
+          },
+        },
+        notes: [
+          "- **Cette route est restreinte aux utilisateurs authentifiés et admin au sein de l'orga**\n" +
+            '- Elle met à jour le numéro étudiant',
+        ],
+        tags: ['api', 'sup-organization-learners'],
       },
     },
   ]);
