+++ api/lib/application/sup-organization-learners/sup-organization-learner-controller.js
@@ -25,4 +25,27 @@ module.exports = {
     }
     return response;
   },
+
+  async updateStudentNumber(request, h) {
+    const payload = request.payload.data.attributes;
+    const organizationId = request.params.id;
+    const studentNumber = payload['student-number'];
+    const organizationLearnerId = request.params.schoolingRegistrationId || request.params.organizationLearnerId;
+
+    await usecases.updateStudentNumber({ organizationLearnerId, studentNumber, organizationId });
+
+    const response = h.response().code(204);
+    if (
+      h.request.path ===
+      `/api/organizations/${organizationId}/schooling-registration-user-associations/${organizationLearnerId}`
+    ) {
+      response
+        .header('Deprecation', 'true')
+        .header(
+          'Link',
+          `/api/organizations/${organizationId}/sup-organization-learners/${organizationLearnerId}; rel="successor-version"`
+        );
+    }
+    return response;
+  },
 };
