+++ api/lib/routes.js
@@ -38,7 +38,6 @@ module.exports = [
   require('./application/progressions'),
   require('./application/saml'),
   require('./application/organization-learner-dependent-users'),
-  require('./application/organization-learner-user-associations'),
   require('./application/organization-learners'),
   require('./application/scorecards'),
   require('./application/sco-organization-learners'),
