+++ api/tests/acceptance/application/sup-organization-learner-controller_test.js
@@ -1,6 +1,7 @@
 const { databaseBuilder, expect, generateValidRequestAuthorizationHeader } = require('../../test-helper');
 
 const createServer = require('../../../server');
+const Membership = require('../../../lib/domain/models/Membership');
 
 describe('Acceptance | Controller | sup-organization-learners', function () {
   let server;
@@ -108,6 +109,96 @@ describe('Acceptance | Controller | sup-organization-learners', function () {
       const response = await server.inject(options);
       // then
       expect(response.statusCode).to.equal(204);
+    });
+  });
+
+  describe('PATCH /api/organizations/id/schooling-registration-user-associations/schoolingRegistrationId', function () {
+    let organizationId;
+    const studentNumber = '54321';
+    let organizationLearnerId;
+    let authorizationToken;
+
+    beforeEach(async function () {
+      organizationId = databaseBuilder.factory.buildOrganization({ isManagingStudents: true, type: 'SUP' }).id;
+
+      const user = databaseBuilder.factory.buildUser();
+      authorizationToken = generateValidRequestAuthorizationHeader(user.id);
+      organizationLearnerId = databaseBuilder.factory.buildOrganizationLearner({ organizationId }).id;
+      databaseBuilder.factory.buildMembership({
+        organizationId,
+        userId: user.id,
+        organizationRole: Membership.roles.ADMIN,
+      });
+      await databaseBuilder.commit();
+    });
+
+    context('Success cases', function () {
+      it('should return an HTTP response with status code 204', async function () {
+        const options = {
+          method: 'PATCH',
+          url: `/api/organizations/${organizationId}/schooling-registration-user-associations/${organizationLearnerId}`,
+          headers: {
+            authorization: authorizationToken,
+          },
+          payload: {
+            data: {
+              attributes: {
+                'student-number': studentNumber,
+              },
+            },
+          },
+        };
+        // when
+        const response = await server.inject(options);
+
+        // then
+        expect(response.statusCode).to.equal(204);
+      });
+    });
+  });
+
+  describe('PATCH /api/organizations/id/sup-organization-learners/organizationLearnerId', function () {
+    let organizationId;
+    const studentNumber = '54321';
+    let organizationLearnerId;
+    let authorizationToken;
+
+    beforeEach(async function () {
+      organizationId = databaseBuilder.factory.buildOrganization({ isManagingStudents: true, type: 'SUP' }).id;
+
+      const user = databaseBuilder.factory.buildUser();
+      authorizationToken = generateValidRequestAuthorizationHeader(user.id);
+      organizationLearnerId = databaseBuilder.factory.buildOrganizationLearner({ organizationId }).id;
+      databaseBuilder.factory.buildMembership({
+        organizationId,
+        userId: user.id,
+        organizationRole: Membership.roles.ADMIN,
+      });
+      await databaseBuilder.commit();
+    });
+
+    context('Success cases', function () {
+      it('should return an HTTP response with status code 204', async function () {
+        const options = {
+          method: 'PATCH',
+          url: `/api/organizations/${organizationId}/sup-organization-learners/${organizationLearnerId}`,
+          headers: {
+            authorization: authorizationToken,
+          },
+          payload: {
+            data: {
+              attributes: {
+                'student-number': studentNumber,
+              },
+            },
+          },
+        };
+        // when
+        const response = await server.inject(options);
+
+        // then
+        expect(response.statusCode).to.equal(204);
+      });
     });
   });
 });
