+++ api/tests/integration/application/sup-organization-learners/index_test.js
@@ -1,7 +1,8 @@
-const { expect, sinon, HttpTestServer } = require('../../../test-helper');
+const { expect, sinon, HttpTestServer, generateValidRequestAuthorizationHeader } = require('../../../test-helper');
 
 const supOrganizationLearnerController = require('../../../../lib/application/sup-organization-learners/sup-organization-learner-controller');
 const moduleUnderTest = require('../../../../lib/application/sup-organization-learners');
+const securityPreHandlers = require('../../../../lib/application/security-pre-handlers');
 
 describe('Integration | Application | Route | sup-organization-learners', function () {
   let httpTestServer;
@@ -10,6 +11,10 @@ describe('Integration | Application | Route | sup-organization-learners', functi
     sinon
       .stub(supOrganizationLearnerController, 'reconcileSupOrganizationLearner')
       .callsFake((request, h) => h.response('ok').code(204));
+    sinon
+      .stub(securityPreHandlers, 'checkUserIsAdminInSUPOrganizationManagingStudents')
+      .callsFake((request, h) => h.response(true));
+    sinon.stub(supOrganizationLearnerController, 'updateStudentNumber').returns('ok');
 
     httpTestServer = new HttpTestServer();
     await httpTestServer.register(moduleUnderTest);
@@ -439,6 +444,244 @@ describe('Integration | Application | Route | sup-organization-learners', functi
 
         // then
         expect(response.statusCode).to.equal(422);
+      });
+    });
+  });
+
+  describe('PATCH /api/organizations/id/schooling-registration-user-associations/schoolingRegistrationId', function () {
+    const method = 'PATCH';
+    const organizationId = 2;
+    const organizationLearnerId = '1234';
+    const userId = 2;
+    let headers;
+
+    beforeEach(function () {
+      headers = {
+        authorization: generateValidRequestAuthorizationHeader(userId),
+      };
+    });
+
+    context('when the user is authenticated', function () {
+      it('should exist', async function () {
+        // given
+        const url = `/api/organizations/${organizationId}/schooling-registration-user-associations/${organizationLearnerId}`;
+        const payload = {
+          data: {
+            attributes: {
+              'student-number': '1234',
+            },
+          },
+        };
+        // when
+        const response = await httpTestServer.request(method, url, payload, null, headers);
+
+        // then
+        expect(response.statusCode).to.equal(200);
+      });
+
+      it('should return a 422 status error when student-number parameter is not a string', async function () {
+        // given
+        const url = `/api/organizations/${organizationId}/schooling-registration-user-associations/${organizationLearnerId}`;
+        const payload = {
+          data: {
+            attributes: {
+              'student-number': 1234,
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload, null, headers);
+
+        // then
+        const responsePayload = JSON.parse(response.payload);
+        expect(response.statusCode).to.equal(422);
+        expect(responsePayload.errors[0].title).to.equal('Unprocessable entity');
+        expect(responsePayload.errors[0].detail).to.equal('Un des champs saisis n’est pas valide.');
+      });
+
+      it('should return a 404 status error when organizationId parameter is not a number', async function () {
+        // given
+        const url = `/api/organizations/FAKE_ORGANIZATION_ID/schooling-registration-user-associations/${organizationLearnerId}`;
+        const payload = {
+          data: {
+            attributes: {
+              'student-number': '1234',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload, null, headers);
+
+        // then
+        const responsePayload = JSON.parse(response.payload);
+        expect(response.statusCode).to.equal(404);
+        expect(responsePayload.errors[0].title).to.equal('Not Found');
+        expect(responsePayload.errors[0].detail).to.equal('Ressource non trouvée');
+      });
+
+      it('should return a 404 status error when studentId parameter is not a number', async function () {
+        // given
+        const url = `/api/organizations/${organizationId}/schooling-registration-user-associations/FAKE_STUDENT_ID`;
+        const payload = {
+          data: {
+            attributes: {
+              'student-number': '1234',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload, null, headers);
+
+        // then
+        const responsePayload = JSON.parse(response.payload);
+        expect(response.statusCode).to.equal(404);
+        expect(responsePayload.errors[0].title).to.equal('Not Found');
+        expect(responsePayload.errors[0].detail).to.equal('Ressource non trouvée');
+      });
+    });
+
+    context('when the user is not authenticated', function () {
+      it('should return an error when the user is not authenticated', async function () {
+        // given
+        securityPreHandlers.checkUserIsAdminInSUPOrganizationManagingStudents.callsFake((request, h) =>
+          h.response().code(403).takeover()
+        );
+        const url = `/api/organizations/${organizationId}/schooling-registration-user-associations/${organizationLearnerId}`;
+        const payload = {
+          data: {
+            attributes: {
+              'student-number': '1234',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload, null, headers);
+
+        // then
+        expect(response.statusCode).to.equal(403);
+      });
+    });
+  });
+
+  describe('PATCH /api/organizations/id/sup-organization-learners/organizationLearnerId', function () {
+    const method = 'PATCH';
+    const organizationId = 2;
+    const organizationLearnerId = '1234';
+    const userId = 2;
+    let headers;
+
+    beforeEach(function () {
+      headers = {
+        authorization: generateValidRequestAuthorizationHeader(userId),
+      };
+    });
+
+    context('when the user is authenticated', function () {
+      it('should exist', async function () {
+        // given
+        const url = `/api/organizations/${organizationId}/sup-organization-learners/${organizationLearnerId}`;
+        const payload = {
+          data: {
+            attributes: {
+              'student-number': '1234',
+            },
+          },
+        };
+        // when
+        const response = await httpTestServer.request(method, url, payload, null, headers);
+
+        // then
+        expect(response.statusCode).to.equal(200);
+      });
+
+      it('should return a 422 status error when student-number parameter is not a string', async function () {
+        // given
+        const url = `/api/organizations/${organizationId}/sup-organization-learners/${organizationLearnerId}`;
+        const payload = {
+          data: {
+            attributes: {
+              'student-number': 1234,
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload, null, headers);
+
+        // then
+        const responsePayload = JSON.parse(response.payload);
+        expect(response.statusCode).to.equal(422);
+        expect(responsePayload.errors[0].title).to.equal('Unprocessable entity');
+        expect(responsePayload.errors[0].detail).to.equal('Un des champs saisis n’est pas valide.');
+      });
+
+      it('should return a 404 status error when organizationId parameter is not a number', async function () {
+        // given
+        const url = `/api/organizations/FAKE_ORGANIZATION_ID/sup-organization-learners/${organizationLearnerId}`;
+        const payload = {
+          data: {
+            attributes: {
+              'student-number': '1234',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload, null, headers);
+
+        // then
+        const responsePayload = JSON.parse(response.payload);
+        expect(response.statusCode).to.equal(404);
+        expect(responsePayload.errors[0].title).to.equal('Not Found');
+        expect(responsePayload.errors[0].detail).to.equal('Ressource non trouvée');
+      });
+
+      it('should return a 404 status error when studentId parameter is not a number', async function () {
+        // given
+        const url = `/api/organizations/${organizationId}/sup-organization-learners/FAKE_STUDENT_ID`;
+        const payload = {
+          data: {
+            attributes: {
+              'student-number': '1234',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload, null, headers);
+
+        // then
+        const responsePayload = JSON.parse(response.payload);
+        expect(response.statusCode).to.equal(404);
+        expect(responsePayload.errors[0].title).to.equal('Not Found');
+        expect(responsePayload.errors[0].detail).to.equal('Ressource non trouvée');
+      });
+    });
+
+    context('when the user is not authenticated', function () {
+      it('should return an error when the user is not authenticated', async function () {
+        // given
+        securityPreHandlers.checkUserIsAdminInSUPOrganizationManagingStudents.callsFake((request, h) =>
+          h.response().code(403).takeover()
+        );
+        const url = `/api/organizations/${organizationId}/sup-organization-learners/${organizationLearnerId}`;
+        const payload = {
+          data: {
+            attributes: {
+              'student-number': '1234',
+            },
+          },
+        };
+
+        // when
+        const response = await httpTestServer.request(method, url, payload, null, headers);
+
+        // then
+        expect(response.statusCode).to.equal(403);
       });
     });
   });
